refactor(dashboard): split KeysPage sections into components/keys/*
Pure move, no behavior change. KeysPage was a 1314-line file holding six
embedded section components; each now lives in its own module:
quota-signals, unified-key, proxy-settings, import-keys, custom-provider,
anthropic, plus shared.tsx for the cross-section helpers (GetKeyLink,
PLATFORMS, status maps, custom-model delete helpers, health types).
KeysPage.tsx shrinks to 482 lines: the page shell, tab state, and the
provider list.
