+++ client/src/components/keys/import-keys-section.tsx
@@ -1,0 +1,247 @@
+import { useState } from 'react'
+import { useMutation, useQueryClient } from '@tanstack/react-query'
+import { apiFetch } from '@/lib/api'
+import { Button } from '@/components/ui/button'
+import { Input } from '@/components/ui/input'
+import { Label } from '@/components/ui/label'
+import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select'
+import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from '@/components/ui/table'
+import type { ImportKey, ImportSelectedResponse, Platform, PreviewKey, PreviewResponse } from '../../../../shared/types'
+import { Upload } from 'lucide-react'
+import { useI18n } from '@/i18n'
+import { PLATFORMS } from './shared'
+
+interface ImportRow extends PreviewKey {
+  selected: boolean
+  platform: Platform | ''
+  visible: boolean
+}
+
+export function ImportKeysSection() {
+  const { t } = useI18n()
+  const queryClient = useQueryClient()
+  const [files, setFiles] = useState<File[]>([])
+  const [rows, setRows] = useState<ImportRow[]>([])
+  const [skipped, setSkipped] = useState<string[]>([])
+  const [result, setResult] = useState<ImportSelectedResponse | null>(null)
+
+  const importablePlatforms = PLATFORMS.filter(p => !p.keyless)
+
+  function platformFromPreview(key: PreviewKey): Platform | '' {
+    return importablePlatforms.some(p => p.value === key.detectedPlatform)
+      ? key.detectedPlatform as Platform
+      : ''
+  }
+
+  const preview = useMutation({
+    meta: { silenceToast: true },
+    mutationFn: async (nextFiles: File[]) => {
+      const formData = new FormData()
+      nextFiles.forEach(file => formData.append('files', file))
+      return apiFetch<PreviewResponse>('/api/keys/preview', { method: 'POST', body: formData })
+    },
+    onSuccess: (data) => {
+      setRows(data.keys.map(key => {
+        const detected = platformFromPreview(key)
+        return {
+          ...key,
+          platform: detected,
+          selected: detected !== '',
+          visible: false,
+        }
+      }))
+      setSkipped(data.skipped)
+      setResult(null)
+    },
+  })
+
+  const importSelected = useMutation({
+    meta: { silenceToast: true },
+    mutationFn: (keys: ImportKey[]) =>
+      apiFetch<ImportSelectedResponse>('/api/keys/import-selected', {
+        method: 'POST',
+        body: JSON.stringify({ keys }),
+      }),
+    onSuccess: (data) => {
+      setResult(data)
+      queryClient.invalidateQueries({ queryKey: ['keys'] })
+      queryClient.invalidateQueries({ queryKey: ['health'] })
+      queryClient.invalidateQueries({ queryKey: ['fallback'] })
+    },
+  })
+
+  const selectedKeys: ImportKey[] = rows
+    .filter(row => row.selected && row.platform && row.keyValue.trim())
+    .map(row => ({
+      keyName: row.keyName,
+      keyValue: row.keyValue,
+      platform: row.platform,
+    }))
+
+  function updateRow(index: number, patch: Partial<ImportRow>) {
+    setRows(prev => prev.map((row, i) => (i === index ? { ...row, ...patch } : row)))
+  }
+
+  function chooseFiles(e: React.ChangeEvent<HTMLInputElement>) {
+    const nextFiles = Array.from(e.target.files ?? [])
+    setFiles(nextFiles)
+    setRows([])
+    setSkipped([])
+    setResult(null)
+    preview.reset()
+    importSelected.reset()
+  }
+
+  return (
+    <section>
+      <div className="mb-3 flex flex-wrap items-center justify-between gap-3">
+        <div>
+          <h2 className="text-sm font-medium">{t('keys.importKeys')}</h2>
+          <p className="text-xs text-muted-foreground mt-0.5">{t('keys.importKeysDescription')}</p>
+        </div>
+        <Button
+          type="button"
+          size="sm"
+          variant="outline"
+          onClick={() => preview.mutate(files)}
+          disabled={files.length === 0 || preview.isPending}
+        >
+          <Upload className="size-3.5" />
+          {preview.isPending ? t('keys.previewing') : t('keys.previewFiles')}
+        </Button>
+      </div>
+
+      <div className="rounded-3xl border bg-card p-4">
+        <div className="flex flex-wrap items-end gap-3">
+          <div className="min-w-[260px] flex-1 space-y-1.5">
+            <Label className="text-xs">{t('keys.importFiles')}</Label>
+            <Input
+              type="file"
+              multiple
+              accept=".env,.json,.jsonc,.md,.txt"
+              onChange={chooseFiles}
+              className="cursor-pointer text-xs file:mr-2"
+            />
+          </div>
+          {files.length > 0 && (
+            <span className="pb-1 text-xs text-muted-foreground">
+              {t('keys.importFileCount', { count: files.length })}
+            </span>
+          )}
+        </div>
+
+        {preview.isError && (
+          <p className="mt-3 text-xs text-destructive">{(preview.error as Error).message}</p>
+        )}
+
+        {rows.length > 0 && (
+          <div className="mt-4 overflow-hidden rounded-2xl border">
+            <Table>
+              <TableHeader>
+                <TableRow>
+                  <TableHead className="w-10">{t('keys.selected')}</TableHead>
+                  <TableHead>{t('keys.provider')}</TableHead>
+                  <TableHead>{t('keys.keyName')}</TableHead>
+                  <TableHead>{t('keys.keyValue')}</TableHead>
+                </TableRow>
+              </TableHeader>
+              <TableBody>
+                {rows.map((row, index) => (
+                  <TableRow key={`${row.keyName}:${index}`}>
+                    <TableCell>
+                      <input
+                        type="checkbox"
+                        checked={row.selected}
+                        onChange={() => updateRow(index, { selected: !row.selected })}
+                        className="size-4 accent-primary"
+                      />
+                    </TableCell>
+                    <TableCell>
+                      <Select
+                        value={row.platform}
+                        onValueChange={(value) => updateRow(index, { platform: value as Platform, selected: true })}
+                      >
+                        <SelectTrigger className="w-[190px]">
+                          <SelectValue placeholder={t('keys.chooseProvider')} />
+                        </SelectTrigger>
+                        <SelectContent>
+                          {importablePlatforms.map(p => (
+                            <SelectItem key={p.value} value={p.value}>{p.label}</SelectItem>
+                          ))}
+                        </SelectContent>
+                      </Select>
+                    </TableCell>
+                    <TableCell>
+                      <Input
+                        value={row.keyName}
+                        onChange={e => updateRow(index, { keyName: e.target.value })}
+                        className="w-[220px] font-mono text-xs"
+                      />
+                    </TableCell>
+                    <TableCell>
+                      <div className="flex min-w-[280px] items-center gap-2">
+                        <Input
+                          type={row.visible ? 'text' : 'password'}
+                          value={row.keyValue}
+                          onChange={e => updateRow(index, { keyValue: e.target.value })}
+                          className="font-mono text-xs"
+                        />
+                        <Button
+                          type="button"
+                          variant="ghost"
+                          size="xs"
+                          onClick={() => updateRow(index, { visible: !row.visible })}
+                        >
+                          {row.visible ? t('common.hide') : t('common.show')}
+                        </Button>
+                      </div>
+                    </TableCell>
+                  </TableRow>
+                ))}
+              </TableBody>
+            </Table>
+          </div>
+        )}
+
+        {rows.length === 0 && preview.isSuccess && (
+          <p className="mt-3 text-xs text-muted-foreground">{t('keys.noPreviewKeys')}</p>
+        )}
+
+        {skipped.length > 0 && (
+          <div className="mt-3 text-xs text-muted-foreground">
+            <span className="font-medium text-foreground">{t('keys.skippedItems')}</span>
+            <span> {skipped.slice(0, 5).join(', ')}</span>
+            {skipped.length > 5 && <span> {t('keys.moreItems', { count: skipped.length - 5 })}</span>}
+          </div>
+        )}
+
+        {rows.length > 0 && (
+          <div className="mt-4 flex flex-wrap items-center gap-3">
+            <Button
+              type="button"
+              size="sm"
+              onClick={() => importSelected.mutate(selectedKeys)}
+              disabled={selectedKeys.length === 0 || importSelected.isPending}
+            >
+              {importSelected.isPending
+                ? t('keys.importing')
+                : t('keys.importSelected', { count: selectedKeys.length })}
+            </Button>
+            {selectedKeys.length === 0 && (
+              <span className="text-xs text-muted-foreground">{t('keys.noImportSelection')}</span>
+            )}
+          </div>
+        )}
+
+        {importSelected.isError && (
+          <p className="mt-3 text-xs text-destructive">{(importSelected.error as Error).message}</p>
+        )}
+        {result && (
+          <p className="mt-3 text-xs text-muted-foreground">
+            {t('keys.importResult', { imported: result.imported, failed: result.errors.length })}
+          </p>
+        )}
+      </div>
+    </section>
+  )
+}
